Minor fixes.
git-svn-id: https://commonkit.googlecode.com/svn/trunk@114 2aaa8e9b-9c4f-0410-8024-4f79e9a52ea3

+++ source/include/MinCOM/Concurrency/Interfaces/IThread.h
@@ -31,7 +31,7 @@ namespace MinCOM
 		
 		static unsigned int GetCores();
 
-		static int GetSelfId();
+        static longlong GetSelfId();
 
         static longlong Time();
